refactor: move the SearchState into its own file and add kdoc
